backend-flask/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
PyJWT==2.8.0
